# [native] Merge some optimizations from new serialization task