--- a/cocos/scene-graph/scene.jsb.ts
+++ b/cocos/scene-graph/scene.jsb.ts
@@ -86,16 +86,17 @@ sceneProto._ctor = function () {
 };
 
 sceneProto._onBatchCreated = function (dontSyncChildPrefab: boolean) {
-    // Don't invoke Node.prototype._onBatchCreated because we refactor Node&BaseNode, BaseNode is empty just for
-    // instanceof check in ts engine. After ts engine removes BaseNode, we could remove BaseNode.h/.cpp too.
     if (this._parent) {
         this._siblingIndex = this._parent.children.indexOf(this);
     }
     //
-    const len = this._children.length;
+    const children = this._children;
+    const len = children.length;
+    let child;
     for (let i = 0; i < len; ++i) {
-        this.children[i]._siblingIndex = i;
-        this._children[i]._onBatchCreated(dontSyncChildPrefab);
+        child = children[i];
+        child._siblingIndex = i;
+        child._onBatchCreated(dontSyncChildPrefab);
     }
 };