--- a/native/cocos/bindings/jswrapper/PrivateObject.h
+++ b/native/cocos/bindings/jswrapper/PrivateObject.h
@@ -196,7 +196,7 @@ inline void inHeap(void *ptr) {
 #endif
 
 template <typename T>
-inline PrivateObjectBase *make_shared_private_object(T *cobj) { // NOLINT
+inline auto *make_shared_private_object(T *cobj) { // NOLINT
     static_assert(!std::is_same<T, void>::value, "void * is not allowed");
 // static_assert(!std::is_pointer_v<T> && !std::is_null_pointer_v<decltype(cobj)>, "bad pointer");
 #if CC_DEBUG
@@ -209,19 +209,19 @@ inline PrivateObjectBase *make_shared_private_object(T *cobj) { // NOLINT
     }
 }
 template <typename T>
-inline PrivateObjectBase *shared_ptr_private_object(std::shared_ptr<T> &&ptr) { // NOLINT
+inline auto *shared_ptr_private_object(std::shared_ptr<T> &&ptr) { // NOLINT
     static_assert(!std::is_base_of<cc::RefCounted, T>::value, "cc::RefCounted is not acceptable for shared_ptr");
     return ccnew SharedPtrPrivateObject<T>(std::forward<std::shared_ptr<T>>(ptr));
 }
 
 template <typename T>
-inline PrivateObjectBase *shared_ptr_private_object(const std::shared_ptr<T> &ptr) { // NOLINT
+inline auto *shared_ptr_private_object(const std::shared_ptr<T> &ptr) { // NOLINT
     static_assert(!std::is_base_of<cc::RefCounted, T>::value, "cc::RefCounted is not acceptable for shared_ptr");
     return ccnew SharedPtrPrivateObject<T>(ptr);
 }
 
 template <typename T>
-inline PrivateObjectBase *rawref_private_object(T *ptr) { // NOLINT
+inline auto *rawref_private_object(T *ptr) { // NOLINT
     // static_assert(false, "always fail");
 //    static_assert(!std::is_base_of<cc::RefCounted, T>::value, "cc::RefCounted is not acceptable for shared_ptr");
 #if CC_DEBUG
@@ -231,12 +231,12 @@ inline PrivateObjectBase *rawref_private_object(T *ptr) { // NOLINT
 }
 
 template <typename T>
-inline PrivateObjectBase *ccintrusive_ptr_private_object(const cc::IntrusivePtr<T> &ptr) { // NOLINT
+inline auto *ccintrusive_ptr_private_object(const cc::IntrusivePtr<T> &ptr) { // NOLINT
     static_assert(std::is_base_of<cc::RefCounted, T>::value, "cc::RefCounted expected!");
     return ccnew CCIntrusivePtrPrivateObject<T>(ptr);
 }
 template <typename T>
-inline PrivateObjectBase *ccintrusive_ptr_private_object(T *cobj) { // NOLINT
+inline auto *ccintrusive_ptr_private_object(T *cobj) { // NOLINT
     static_assert(std::is_base_of<cc::RefCounted, T>::value, "cc::RefCounted expected!");
     return ccnew CCIntrusivePtrPrivateObject<T>(cc::IntrusivePtr<T>(cobj));
 }