… native/CMakeLists.txt
… native/cocos/bindings/jswrapper/v8/Class.h
… native/cocos/bindings/jswrapper/v8/HelperMacros.cpp
… native/cocos/bindings/jswrapper/v8/HelperMacros.h
… native/cocos/bindings/jswrapper/v8/Object.h
… native/cocos/bindings/jswrapper/v8/Utils.cpp
… native/cocos/bindings/manual/jsb_conversions.h
… native/cocos/bindings/manual/jsb_global.h
… native/cocos/bindings/sebind/class_v8.h
… native/cocos/core/TypedArray.h
… native/cocos/core/data/Object.h
… native/cocos/core/scene-graph/Node.cpp
… native/cocos/core/scene-graph/Node.h
… native/tests/unit-test/src/base_has_member_function_test.cpp
… native/tools/swig-config/cocos.i
… native/tools/swig-config/scene.i